provider/sl: make the code more sparse
